Can draw a primitive point in space
Former-commit-id: 5ae0bac391bf78a6c78f7ea1a205e8c50d341d5c [formerly 16dafb6bde4262d8b8d683c44adc5e61d085880b [formerly 40d7f7e83cf1b616d3b5708748f17979a3f965b7]]
Former-commit-id: d976d619ffeaf0eea7c9e1dfef5aec4f97a8c87f
Former-commit-id: 884c05dd6078ea51b7cf85a516264d1816d62348

+++ BrickwareCore/src/Primitive.cpp
@@ -119,7 +119,7 @@ void Primitive::createBuffersGL()
 
 	glGenBuffers(1, &ibo);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, ibo);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, pointCount * sizeof(unsigned int), pointIndices, GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, pointCount * sizeof(unsigned short), pointIndices, GL_STATIC_DRAW);
 
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
